Apply suggestions from code review
Apply suggestions to prevent implicit type conversion.

+++ portable/MemMang/heap_4.c
@@ -618,9 +618,9 @@ void vPortHeapResetState( void )
 {
     pxEnd = NULL;
 
-    xFreeBytesRemaining = 0U;
-    xMinimumEverFreeBytesRemaining = 0U;
-    xNumberOfSuccessfulAllocations = 0;
-    xNumberOfSuccessfulFrees = 0;
+    xFreeBytesRemaining = ( size_t) 0U;
+    xMinimumEverFreeBytesRemaining = ( size_t )0U;
+    xNumberOfSuccessfulAllocations = ( size_t ) 0U;
+    xNumberOfSuccessfulFrees = ( size_t ) 0U;
 }
 /*-----------------------------------------------------------*/
